+++ assets/aml-benchmark/components/src/prompt_crafter/package/prompt_factory.py
@@ -330,6 +330,11 @@ class ChatPromptFactory(PromptFactory):
             # NOTE: user role is currently prefered over system role
             # though this may change in the future as models are updated
             messages = [{"role": role, "content": affix}]
+
+        # Handling the case where Json.loads load a string with only number as intger.
+        if isinstance(messages, int):
+            messages = [{"role": role, "content": affix}]
+
         messages = self._make_list(messages)
         self._validate_messages(messages)
         return messages