+++ assets/training/model_evaluation/src/download_dependencies.py
@@ -5,14 +5,14 @@
 
 from mlflow.pyfunc import _get_model_dependencies
 from utils import ArgumentParser
-from exceptions import swallow_all_exceptions
-from logging_utilities import custom_dimensions, get_logger, log_traceback
+from logging_utilities import (
+    custom_dimensions, get_logger, log_traceback, swallow_all_exceptions, get_azureml_exception
+)
 from azureml.telemetry.activity import log_activity
-import constants
-# import traceback
 from error_definitions import DownloadDependenciesError
 from exceptions import ModelValidationException
-from azureml._common._error_definition.azureml_error import AzureMLError
+from validation import _validate_model
+from constants import ArgumentLiterals, TelemetryConstants
 
 try:
     from pip import main as pipmain
@@ -20,8 +20,8 @@ except ImportError:
     from pip._internal import main as pipmain
 REQUIREMENTS_FILE = "./requirements.txt"
 
+custom_dimensions.app_name = TelemetryConstants.DOWNLOAD_MODEL_DEPENDENCIES
 logger = get_logger(name=__name__)
-custom_dimensions.app_name = constants.TelemetryConstants.COMPONENT_NAME
 custom_dims_dict = vars(custom_dimensions)
 
 
@@ -29,11 +29,14 @@ custom_dims_dict = vars(custom_dimensions)
 def main():
     """Download and Install mlflow model dependencies."""
     parser = ArgumentParser()
-    parser.add_argument("--mlflow-model", type=str, dest="mlflow_model", required=True)
+    parser.add_argument("--mlflow-model", type=str, dest=ArgumentLiterals.MLFLOW_MODEL, required=True)
 
     args, _ = parser.parse_known_args()
+    args = vars(args)
 
-    reqs_file = _get_model_dependencies(args.mlflow_model, "pip")
+    _validate_model(args)
+
+    reqs_file = _get_model_dependencies(args[ArgumentLiterals.MLFLOW_MODEL], "pip")
 
     ignore_deps = [
         "mlflow",
@@ -43,7 +46,7 @@ def main():
         "azureml-metrics"
     ]
     install_deps = []
-    with log_activity(logger, constants.TelemetryConstants.DOWNLOAD_MODEL_DEPENDENCIES,
+    with log_activity(logger, TelemetryConstants.DOWNLOAD_MODEL_DEPENDENCIES,
                       custom_dimensions=custom_dims_dict):
         with open(reqs_file, "r") as f:
             for line in f.readlines():
@@ -55,12 +58,11 @@ def main():
             try:
                 pipmain(["install"] + install_deps)
             except Exception as e:
-                exception = ModelValidationException._with_error(
-                    AzureMLError.create(DownloadDependenciesError, error=repr(e),
-                                        message_kwargs={"dependencies": ' '.join(install_deps)}),
-                    inner_exception=e
-                )
+                message_kwargs = {"dependencies": ' '.join(install_deps)}
+                exception = get_azureml_exception(ModelValidationException, DownloadDependenciesError, e,
+                                                  error=repr(e), **message_kwargs)
                 log_traceback(exception, logger)
+                raise exception
 
 
 if __name__ == "__main__":