+++ assets/training/model_management/src/azureml/model/mgmt/utils/common_utils.py
@@ -329,3 +329,35 @@ def get_git_lfs_blob_size_in_kb(git_dir: Path) -> int:
         return int(stdout)
     except Exception as e:
         raise AzureMLException._with_error(AzureMLError.create(GenericRunCMDError, error=e))
+
+
+class MlflowMetaConstants:
+    """Mlflow consants."""
+
+    IS_FINETUNED_MODEL = "is_finetuned_model"
+    IS_ACFT_MODEL = "is_acft_model"
+    BASE_MODEL_NAME = "base_model_name"
+
+
+def fetch_mlflow_acft_metadata(
+        is_finetuned_model: bool = False,
+        base_model_name: str = None) -> dict:
+    """Fetch metadata to be dumped in MlFlow MlModel File.
+
+    :param is_finetuned_model: whether the model is finetuned one or base model
+    :type is_finetuned_model: bool
+    :param is_acft_model: whether the model using acft packages
+    :type is_acft_model: bool
+    :param base_model_name: name of the model
+    :type base_model_name: str
+
+    :return: metadata
+    :rtype: dict
+    """
+    metadata = {
+        MlflowMetaConstants.IS_FINETUNED_MODEL: is_finetuned_model,
+        MlflowMetaConstants.IS_ACFT_MODEL: True,
+        MlflowMetaConstants.BASE_MODEL_NAME: base_model_name
+    }
+
+    return metadata