+++ scripts/azureml-assets/CHANGELOG.md
@@ -3,6 +3,14 @@
 
 ### 🐛 Bugs Fixed
 
+## 1.16.14
+### 🐛 Bugs Fixed
+- [#1574](https://github.com/Azure/azureml-assets/pull/1574) Fix fetching validated model assets
+
+## 1.16.13
+### 🚀 New Features
+- [#1569](https://github.com/Azure/azureml-assets/pull/1569) Add model validations
+
 ## 1.16.12 (2023-10-12)
 ### 🐛 Bugs Fixed
 - [#1458](https://github.com/Azure/azureml-assets/pull/1458) Update schema for prompt and benchmark assets